+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## _Under development_
 
+- Scan
+    - Monitoring
+        - Fixed bug causing the error log not to appear when there are errors.
+
 ## Version 0.4.1 _(July 06, 2013)_
 
 - Added welcome screen after first sign-in.